SystemRecomandation/requirements.txt
arrow==1.4.0
asgiref==3.10.0
certifi==2025.11.12
cffi==2.0.0
circuitbreaker==2.1.3
click==8.0.4
colorama==0.4.6
cryptography==45.0.7
Django==5.2.8
django-debug-toolbar==6.1.0
djangorestframework==3.16.1
jmespath==0.10.0
joblib==1.5.2
numpy==2.3.5
oci==2.164.0
oci-cli==3.71.0
pandas==2.3.3
pillow==12.0.0
prompt-toolkit==3.0.43
pycparser==2.23
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
terminaltables==3.1.10
threadpoolctl==3.6.0
transliterate==1.10.2
typing_extensions==4.15.0
tzdata==2025.2
wcwidth==0.2.14